requirements.txt
Flask
bcrypt
PyOTP
python-dotenv
gunicorn